intersection Tubemask and blackpixel implemented
during the area compensation process it is no longer possible that the analysed area will include area outside of the tube when increasing the black circle radius. also a minor change for the upper radius candidate limit is implemented. also the issue that pycharm could not find the right interpreter is fixed